requirements.txt
Pillow==9.0.1
Pillow==10.0.0
streamlit==1.23.1
